complete signal process
