+++ minishell/initial/initialize_meta_token.c
@@ -6,7 +6,7 @@
 /*   By: jahong <jahong@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 15:18:58 by jahong            #+#    #+#             */
-/*   Updated: 2025/03/18 15:26:15 by jahong           ###   ########.fr       */
+/*   Updated: 2025/03/18 20:49:05 by jahong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
