+++ minishell/resource_free/free_resources.c
@@ -6,11 +6,12 @@
 /*   By: jahong <jahong@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/14 10:00:29 by jahong            #+#    #+#             */
-/*   Updated: 2025/03/18 16:44:23 by jahong           ###   ########.fr       */
+/*   Updated: 2025/03/19 12:39:13 by jahong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
+#include "../syntax/syntax.h"
 
 void	free_resources(t_data *meta, int **pipes, char *path, int exit_n)
 {
@@ -28,4 +29,17 @@ void	free_resources(t_data *meta, int **pipes, char *path, int exit_n)
 		g_ws = exit_n;
 		exit(exit_n);
 	}
+}
+
+int	rutin_free(t_data *meta, char *str)
+{
+	if (meta->exec_cmd)
+	{
+		//printf_exec_commads(meta->exec_cmd);
+		free_exec_linked_list(meta->exec_cmd);
+		meta->exec_cmd = NULL;
+	}
+	meta->tokens = free_t_list(meta->tokens);
+	add_history_and_free(&str);
+	return (0);
 }
